refactor(commands): split command service into basic infrastructure, command implementations and outside api
