+++ opengs_maptool/services/commands/__init__.py
@@ -1,0 +1,3 @@
+from . import system_commands
+from . import project_commands
+from . import import_commands
